+++ src/arp_resources/arp_reach/pcd_processing/launch/pcd_service_launch.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0"?>
+<!--The Launch file for the service pcd processor node!-->
+<launch>
+    <node pkg="pcd_processing" exec="PoseArrayToPCDService.py" output="screen"></node>
+</launch>
