+++ src/arp_resources/arp_reach/pcd_processing/package.xml
@@ -9,18 +9,19 @@
  <maintainer email="ntchmura@gmail.com">Natalie Chmura</maintainer>
  <license>Apache License 2.0</license>
 
-<buildtool_depend>ament_python</buildtool_depend>
+ <buildtool_depend>ament_cmake</buildtool_depend>
+ <buildtool_depend>ament_cmake_python</buildtool_depend>
 
-  <exec_depend>rclpy</exec_depend>
+
+ <exec_depend>rclpy</exec_depend>
 
  <exec_depend>std_msgs</exec_depend>
  <exec_depend>geometry_msgs</exec_depend>
  <exec_depend>arp_msgs</exec_depend>
 
-
  <exec_depend>python3</exec_depend>
 
  <export>
-    <build_type>ament_python</build_type>
+    <build_type>ament_cmake</build_type>
  </export>
 </package>
