--- src/arp_resources/arp_reach/pcd_processing/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/pcd_processing
-[install]
-install_scripts=$base/lib/pcd_processing
